chore(cleanup): general cleanup
