+++ webapp/src/app/files/file.module.ts
@@ -4,11 +4,12 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 
 import { FileUploadModule } from 'ng2-file-upload';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { UiModule } from '../ui/ui.module';
 import { FileComponent } from './file.component';
 import { FileFormComponent } from './form/file-form.component';
-import { UploadFormComponent } from './form/upload-form.component';
+import { UploadFormComponent } from './upload/upload-form.component';
 import { ContainerListComponent } from './list/container-list.component';
 
 import { FileService } from './file.service';
@@ -19,7 +20,8 @@ import { FileService } from './file.service';
     FormsModule,
     UiModule,
     FileUploadModule,
-    RouterModule
+    RouterModule,
+    NgbModule
   ],
   declarations: [
     FileComponent,
