+++ webapp/src/app/shared/sdk/guards/user-logged.ts
@@ -12,7 +12,7 @@ export class UserLoggedGuard implements CanActivate {
   constructor(private store: Store<any>) { }
 
   public canActivate(): Observable<boolean> {
-    return this.store.let(getLoopbackAuthUserId)
+    return this.store.let(getLoopbackAuthUserId())
       .flatMap((userId) => {
         return userId ? of(true) : of(false);
       });
