+++ webapp/src/app/shared/sdk/reducers/auth.ts
@@ -2,8 +2,7 @@
 import { Observable } from 'rxjs/Observable';
 import { createSelector } from 'reselect';
 import { Action } from '@ngrx/store';
-import { SDKToken } from '../models/BaseModels';
-import { Account } from '../models/Account';
+import { SDKToken, Account } from '../models';
 import { LoopbackAuthActionTypes } from '../actions/auth';
 import { UserActionTypes } from '../actions/user';
