+++ webapp/src/app/ui/card/card.component.ts
@@ -14,17 +14,17 @@ import { NavItem } from '../ui.service';
         <div class="card-header bg-primary" *ngIf="title">
           <div class="row align-items-center">
             <div class="col-12 col-md-7">
-              <h5 class="card-title text-uppercase text-white mb-0">
+              <h4 class="card-title text-uppercase text-white mb-0">
                 <i *ngIf="icon" [class]="'fa fa-fw fa-' + icon"></i> {{ title }}
-              </h5>
+              </h4>
               <hr *ngIf="createButton || subTitle">
             </div>
             <div class="col-12 col-md-5">
-            <p *ngIf="subTitle" class="card-subtitle lead float-right">{{ subTitle }}</p>
               <button *ngIf="createButton" [class]="createButton.class" (click)="action.emit($event)">
                 <i [class]="'fa fa-fw fa-lg fa-' + createButton.icon"></i>
                  {{ createButton.text }}
               </button>
+              <p *ngIf="subTitle" class="card-subtitle lead float-right">{{ subTitle }}</p>
             </div>
             <div *ngIf="nav" class="col-12">
               <ul class="nav nav-tabs card-header-tabs">
