+++ webapp/src/app/ui/styles/_custom.scss
@@ -13,6 +13,10 @@ h5 {
   // text-shadow: 1px 1px 2px rgba(0,0,0,0.3);
 }
 
+.dropdown-container {
+  width: 100%;
+}
+
 .breadcrumb {
   background: #fff;
   font-size: 13px;
